LayoutTests/storage/indexeddb/basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test IndexedDB's basics.");
if (window.layoutTestController)
    layoutTestController.waitUntilDone();

function test()
{
    shouldBeTrue("'webkitIndexedDB' in window");
    shouldBeFalse("webkitIndexedDB == null");

    // FIXME: Verify other IndexedDatabaseRequest constructors, once they're implemented.

    result = evalAndLog("webkitIndexedDB.open('name', 'description')");
    verifyResult(result);
    result.onsuccess = openCallback;
    result.onerror = unexpectedErrorCallback;
}

function openCallback()
{
    verifySuccessEvent(event);
    try {
        debug("webkitIndexedDB.open('name');");
        webkitIndexedDB.open('name');
        testFailed("Calling IDBFactory::open without a description should have thrown.");
    } catch (err) {
        testPassed("Exception thrown.");
    }
    try {
        debug("webkitIndexedDB.open('name', null);");
        webkitIndexedDB.open('name', null);
        testFailed("Calling IDBFactory::open with a null description should have thrown.");
    } catch (err) {
        testPassed("Exception thrown.");
    }
    try {
        debug("webkitIndexedDB.open('name', undefined);");
        webkitIndexedDB.open('name', undefined);
        testFailed("Calling IDBFactory::open with an undefined description should have thrown.");
    } catch (err) {
        testPassed("Exception thrown.");
    }
    done();
}


test();

var successfullyParsed = true;

</script>
</body>
</html>
